fix(embed): address founder review on spelunk-embed extraction (oss^82)
Three fixes from the PR #515 review:

- Restore the macOS-only cfg gate on `libc`. The pre-extraction
  spelunk-server Cargo.toml pinned `libc` only inside a
  `[target.'cfg(target_os = "macos")'.dependencies]` block; the refactor
  had promoted it to an unconditional `[workspace.dependencies]` entry.
  Cargo has no way to gate workspace.dependencies by target, so libc is
  now pinned directly inside spelunk-embed's own macOS-gated target
  table instead (its only consumer), matching the original layout.

- Drop the duplicate `tokio` dev-dependency in spelunk-embed/Cargo.toml.
  The crate has no separate integration-test crate (only an inline
  `#[cfg(test)] mod tests`), so the existing `[dependencies]` entry
  already covers both production and test code; the dev-dependency
  entry was fully redundant.

- Move the Hugging Face Hub download/quantize path out of spelunk-embed
  and into a new `embed_hub` module in spelunk-server. spelunk-embed now
  exposes only `load_from_path` (pure local-file load, zero network) and
  no longer depends on `hf-hub`, `dirs`, or their transitive deps.
  spelunk-server's embed_hub resolves the model artifacts via hf-hub and
  then calls spelunk-embed's `load_from_path`. Verified with
  `cargo tree -p spelunk-embed`: no hf-hub anywhere in that crate's
  dependency graph. Updated docs/architecture.md, docs/third-party-models.md,
  and CLAUDE.md to reflect the new module split.
