+++ crates/spelunk-embed/Cargo.toml
@@ -12,8 +12,11 @@ path = "src/lib.rs"
 
 [features]
 default = ["embed-native"]
-# embed-native: bundle the F2LLM-v2-330M Qwen3 embedder via candle (CPU).
-embed-native = ["dep:candle-core", "dep:candle-nn", "dep:candle-transformers", "dep:hf-hub", "dep:tokenizers"]
+# embed-native: bundle the F2LLM-v2-330M Qwen3 embedder via candle (CPU). No
+# download dependency here (no hf-hub) — this crate only loads from local
+# files (`load_from_path`); the Hugging Face Hub acquisition path lives in
+# spelunk-server's `embed_hub` module.
+embed-native = ["dep:candle-core", "dep:candle-nn", "dep:candle-transformers", "dep:tokenizers"]
 # metal: enable Metal GPU acceleration on macOS. Add this when building the macOS binary.
 metal = ["candle-core/metal", "candle-nn/metal", "candle-transformers/metal"]
 
@@ -24,20 +27,19 @@ async-trait = { workspace = true }
 serde_json = { workspace = true }
 tokio = { workspace = true }
 tracing = { workspace = true }
-dirs = { workspace = true }
 candle-core = { workspace = true, optional = true }
 candle-nn = { workspace = true, optional = true }
 candle-transformers = { workspace = true, optional = true }
-hf-hub = { workspace = true, optional = true }
 tokenizers = { workspace = true, optional = true }
 
 # macOS-only: `hw.memsize` sysctl in `total_system_ram` (Linux reads
-# /proc/meminfo, no FFI).
+# /proc/meminfo, no FFI). Not in [workspace.dependencies] since spelunk-embed
+# is the only consumer; a bare version here keeps the cfg gate unambiguous
+# (Cargo has no way to gate workspace.dependencies itself by target).
 [target.'cfg(target_os = "macos")'.dependencies]
-libc = { workspace = true }
+libc = "0.2"
 
 [dev-dependencies]
 tempfile = { workspace = true }
 pretty_assertions = { workspace = true }
 serial_test = { workspace = true }
-tokio = { workspace = true }
